+++ web/src/App.tsx
@@ -1,8 +1,8 @@
 import AppHeader from "./components/header/header";
-import CommitteeHub from "./components/committee/committeePage";
+import CommitteeHub from "./components/committee";
 import { Route, Routes } from "react-router-dom";
 import { HeaderProvider } from "./contexts/headerContext";
-import SelectCommittee from "./components/committee/components/selectCommittee/selectCommittee";
+import SelectCommittee from "./components/committee/components/selectCommittee";
 import SelectDelegation from "./components/delegation/components/selectDelegation/selectDelegation";
 import DelegationPage from "./components/delegation/delegationPage";
 import { Box } from "@mui/material";
