--- web/src/components/committee/components/attendance/attendance.css
@@ -1,23 +1,0 @@
-.list {
-  padding: 10px;
-}
-
-.footer {
-  display: flex;
-  padding: 10px;
-}
-
-.footer .addButton {
-  flex: 1;
-  display: flex;
-  align-items: center;
-}
-
-.footer .selectCount {
-  display: flex;
-  align-items: center;
-}
-
-.footer .selectCount .select-label {
-  padding: 10px;
-}
