+++ web/src/components/committee/components/speakers-list/index.tsx
