+++ web/src/components/committee/components/workingPapers/workingPapers.css
@@ -1,11 +1,3 @@
-/* Styles for the table */
-table {
-  width: 100%;
-  border-collapse: collapse;
-  font-size: 16px;
-  text-align: left;
-}
-
 th,
 td {
   padding: 12px;
